+++ packages/vite-plugin-cloudflare/playground/node-env/package.json
@@ -1,13 +1,13 @@
 {
-	"name": "@playground/react-dom-server",
+	"name": "@playground/node-env",
 	"private": true,
 	"type": "module",
 	"scripts": {
 		"build": "vite build",
-		"build:with-nodejs-compat": "vite build -c vite.config.with-nodejs-compat.ts",
+		"build:nodejs-compat": "vite build -c vite.config.nodejs-compat.ts",
 		"check:types": "tsc --build",
 		"dev": "vite dev",
-		"dev:with-nodejs-compat": "vite dev -c vite.config.with-nodejs-compat.ts",
+		"dev:nodejs-compat": "vite dev -c vite.config.nodejs-compat.ts",
 		"preview": "vite preview"
 	},
 	"dependencies": {
