+++ packages/vite-plugin-cloudflare/playground/node-env/vite.config.nodejs-compat.ts
@@ -6,7 +6,7 @@ export default defineConfig({
 		cloudflare({
 			inspectorPort: false,
 			persistState: false,
-			configPath: "wrangler.with-nodejs-compat.jsonc",
+			configPath: "wrangler.nodejs-compat.jsonc",
 		}),
 	],
 });
