--- packages/vite-plugin-cloudflare/playground/react-dom-server/tsconfig.worker.json
@@ -1,4 +1,0 @@
-{
-	"extends": ["@cloudflare/workers-tsconfig/worker.json"],
-	"include": ["src"]
-}
